requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
billiard==3.3.0.23
celery==3.1.26.post2
Django==3.2.7
django-appconf==1.0.4
django-celery==3.3.1
django-imagekit==4.0.2
django-rest==0.8.7
djangorestframework==3.12.2
kombu==3.0.37
pilkit==2.0
Pillow==8.3.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
